refactor(image-annotation): :recycle: move the delete crop button
