+++ src/screens/annotation/image-annotation-screen/components/CropScrollView.tsx
@@ -6,7 +6,10 @@ import { useAppDispatch, useAppSelector } from '../../../../stores/hooks';
 import { useCurrentSelectedCropContext } from '../context/CurrentSelectedCropContext';
 import { useDisplayedImageSizeContext } from '../context/DisplayedImageSizeContext';
 import { useTrueImageSizeContext } from '../context/TrueImageSizeContext';
-import { setCurrentAnnotatedImagePixels } from '../current-annotated-image';
+import {
+  currentAnnotatedImageRemoveCrop,
+  setCurrentAnnotatedImagePixels,
+} from '../current-annotated-image';
 import { Crop, Pixel, Point } from '../types/image-annotation-types';
 import { getAllPointsInPath } from '../utils/pixels-utils';
 import CropContainer from './CropContainer';
@@ -62,6 +65,28 @@ const CropScrollView = () => {
    */
   const selectCrop = (index: number) => {
     setCurrentSelectedCrop(index);
+  };
+
+  /**
+   * Removes the crop from the redux store
+   */
+  const deleteCrop = () => {
+    if (currentSelectedCrop !== undefined && trueImageSize) {
+      const pixelsCopy: Pixel[] = cloneDeep(currentImage.imagePixels);
+
+      getAllPointsInPath(
+        currentImage.imageCrops[currentSelectedCrop].cropPath,
+        trueImageSize.width,
+      ).forEach((index: number) => {
+        // For every point, sets the annotation of the corresponding pixel to the one of the crop, if the pixel is not white (background)
+        const pixel: Pixel = pixelsCopy[index];
+        pixel.annotation = undefined;
+      });
+
+      dispatch(currentAnnotatedImageRemoveCrop(currentSelectedCrop));
+      dispatch(setCurrentAnnotatedImagePixels(pixelsCopy));
+      setCurrentSelectedCrop(undefined);
+    }
   };
 
   /**
@@ -143,7 +168,7 @@ const CropScrollView = () => {
           />
         ))}
       </ScrollView>
-      <CropContainerButtons annotate={annotate} />
+      <CropContainerButtons annotate={annotate} deleteCrop={deleteCrop} />
     </View>
   );
 };
